feat(cli): extract cli to own file
